Added implementation for all Spark jobs.
Just add your SQL query where it says add SQL query, and be sure to
change any data types to the matching data type in the Java file if need
be. The RedFlags is a bit different with what tables need to be used, so
we will discuss that at a later time.